ref: Move `ProgressBarMode` code to `utils`
This code is more closely related to the utils module's progress bar code than to the api code.

Ref #2008
