+++ src/utils/progress.rs
@@ -1,3 +1,4 @@
+use parking_lot::RwLock;
 use std::env;
 use std::ops::Deref;
 use std::sync::Arc;
@@ -70,5 +71,30 @@ impl Deref for ProgressBar {
 
     fn deref(&self) -> &Self::Target {
         &self.inner
+    }
+}
+
+#[derive(Clone)]
+pub enum ProgressBarMode {
+    Disabled,
+    Request,
+    Response,
+    Shared((Arc<ProgressBar>, u64, usize, Arc<RwLock<Vec<u64>>>)),
+}
+
+impl ProgressBarMode {
+    /// Returns if progress bars are generally enabled.
+    pub fn active(&self) -> bool {
+        !matches!(*self, ProgressBarMode::Disabled)
+    }
+
+    /// Returns whether a progress bar should be displayed during upload.
+    pub fn request(&self) -> bool {
+        matches!(*self, ProgressBarMode::Request)
+    }
+
+    /// Returns whether a progress bar should be displayed during download.
+    pub fn response(&self) -> bool {
+        matches!(*self, ProgressBarMode::Response)
     }
 }
